Build and release shell.efi
The UEFI Shell application which is embedded in the edk2/ovmf firmware
is useful as a stand-alone app and can be installed as an fallback boot
application (EFI/BOOT/BOOTX64.EFI) and coupled with a startup.nsh script
to programatically run EFI applications.

Since we already build EFI applications for showpcr, also copy out the
Shell.efi application and publish it.